Divided structs into more modules.
