+++ src/base.rs
@@ -1,0 +1,8 @@
+use holder::*;
+
+pub trait Watchable {
+    fn watch(&self) -> String;
+}
+
+key_gen!(RoomKey);
+key_gen!(ActorKey);
